requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image
matplotlib>=3.8.0
tqdm>=4.66.0
diffusers>=0.21.0
accelerate
transformers>=4.35.0
huggingface_hub>=0.17.0
facenet-pytorch
torch>=2.1.0
transformers>=4.35.0
huggingface_hub>=0.17.0
omegaconf
fvcore